Massive GUI Overhaul.
Massively overhauled the GUI to make it look presentable along with cleaning up some parts of the code and popups. Also added a background image. Fixed the search feature to work without a backend too. The projects popup seems to have broken in converting to canvas, this error should be fixed soon.